requirements.txt
torch
torchvision
ultralytics
open_clip_torch
Pillow
numpy